StageSpaceUI: put ablation command in own file
